+++ Tests/Fixtures/InvokableWithInputTestCommand.php
@@ -13,7 +13,7 @@ namespace Symfony\Component\Console\Tests\Fixtures;
 
 use Symfony\Component\Console\Attribute\Argument;
 use Symfony\Component\Console\Attribute\AsCommand;
-use Symfony\Component\Console\Attribute\Input;
+use Symfony\Component\Console\Attribute\MapInput;
 use Symfony\Component\Console\Attribute\Option;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Style\SymfonyStyle;
@@ -21,7 +21,7 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 #[AsCommand('invokable:input:test')]
 class InvokableWithInputTestCommand
 {
-    public function __invoke(SymfonyStyle $io, #[Input] UserDto $user): int
+    public function __invoke(SymfonyStyle $io, #[MapInput] UserDto $user): int
     {
         $io->writeln($user->name);
         $io->writeln($user->email);
@@ -47,7 +47,7 @@ final class UserDto
     #[Argument]
     public string $password;
 
-    #[Input]
+    #[MapInput]
     public UserGroupDto $group;
 
     #[Option]